style: refactor youtube package to incorporate common interfaces
